beamer: change agent structure
