+++ beamer/agent/typing.py
