+++ beamer/tests/agent/unit/test_challenge.py
@@ -1,7 +1,8 @@
 import pytest
 from web3.types import Wei
 
-from beamer.chain import process_claims
+from beamer.agent.chain import process_claims
+from beamer.agent.typing import FillId, Termination
 from beamer.tests.agent.unit.utils import (
     ADDRESS1,
     CLAIMER_STAKE,
@@ -13,7 +14,6 @@ from beamer.tests.agent.unit.utils import (
 )
 from beamer.tests.agent.utils import make_address
 from beamer.tests.constants import FILL_ID
-from beamer.typing import FillId, Termination
 
 
 @pytest.mark.parametrize("fill_id", [FILL_ID, FillId(b"cafebabe")])
