+++ beamer/tests/contracts/test_l1_resolution.py
@@ -4,10 +4,10 @@ from brownie import web3
 from brownie.convert.datatypes import HexString
 from web3.constants import ADDRESS_ZERO
 
+from beamer.agent.typing import RequestId
 from beamer.tests.agent.utils import make_address
 from beamer.tests.constants import FILL_ID, FILL_ID_EMPTY, RM_R_FIELD_FILL_ID, RM_R_FIELD_FILLER
 from beamer.tests.util import alloc_accounts, alloc_whitelisted_accounts, create_request_id
-from beamer.typing import RequestId
 
 
 @pytest.mark.parametrize("amount", [100, 99, 101])
